requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
requests==2.31.0
FastAPI==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
